Separate TS types to be managed manually
