+++ .changeset/fair-carpets-beg.md
@@ -1,0 +1,5 @@
+---
+"openapi-fetch": patch
+---
+
+Separate TS types to be managed manually
