+++ packages/openapi-fetch/test/fixtures/api.d.ts
