+++ packages/openapi-fetch/test/fixtures/api.yaml
