+++ packages/openapi-fetch/test/fixtures/v7-beta.d.ts
