+++ packages/openapi-fetch/tsconfig.json
@@ -7,6 +7,7 @@
     "lib": ["ESNext", "DOM"],
     "module": "NodeNext",
     "moduleResolution": "NodeNext",
+    "noUncheckedIndexedAccess": true,
     "outDir": "dist",
     "skipLibCheck": true,
     "strict": true,
