+++ packages/openapi-fetch/vitest.config.ts
@@ -1,7 +1,8 @@
-import { defineConfig } from 'vitest/config';
+import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
+    environment: "jsdom",
     globals: true,
   },
 });
