+++ packages/openapi-typescript/.eslintignore
@@ -1,0 +1,1 @@
+test/fixtures
