+++ packages/openapi-typescript/package.json
@@ -72,7 +72,7 @@
     "@types/node": "^20.8.0",
     "degit": "^2.8.4",
     "del-cli": "^5.1.0",
-    "esbuild": "^0.19.4",
+    "esbuild": "^0.19.5",
     "execa": "^7.2.0",
     "vite": "^4.4.9",
     "vite-node": "^0.34.6",
